WebSocket ping-pongs the connection
There are some circumstances where the connection my just disappear (such as the computer going to sleep). When it wakes back up, the socket is not restored. The WebSocket wrapper has a ~5 second timer it uses to ping the connection to make sure it's still active. If it doesn't get a pong within 5 seconds, it shuts down the socket, which ultimately results in a state change and a re-initialization of the entire connection.